+++ src/platform.js
@@ -1,6 +1,6 @@
 'use strict'
 
-const asar = require('asar')
+const asar = require('@electron/asar')
 const crypto = require('crypto')
 const debug = require('debug')('electronite-packager')
 const fs = require('fs-extra')
@@ -61,16 +61,35 @@ class App {
     if (this.opts.tmpdir === false) {
       return common.generateFinalPath(this.opts)
     } else {
-      return path.join(
-        common.baseTempDir(this.opts),
-        `${this.opts.platform}-${this.opts.arch}`,
-        common.generateFinalBasename(this.opts)
-      )
+      if (!this.cachedStagingPath) {
+        const parentDir = path.join(
+          common.baseTempDir(this.opts),
+          `${this.opts.platform}-${this.opts.arch}`
+        )
+        fs.mkdirpSync(parentDir)
+        this.cachedStagingPath = fs.mkdtempSync(path.join(parentDir, `${common.generateFinalBasename(this.opts)}-`))
+      }
+      return this.cachedStagingPath
     }
   }
 
   get appAsarPath () {
     return path.join(this.originalResourcesDir, 'app.asar')
+  }
+
+  get commonHookArgs () {
+    return [
+      this.opts.electronVersion,
+      this.opts.platform,
+      this.opts.arch
+    ]
+  }
+
+  get hookArgsWithOriginalResourcesAppDir () {
+    return [
+      this.originalResourcesAppDir,
+      ...this.commonHookArgs
+    ]
   }
 
   async relativeRename (basePath, oldName, newName) {
@@ -107,6 +126,8 @@ class App {
     } else {
       await this.buildApp()
     }
+
+    await hooks.promisifyHooks(this.opts.afterInitialize, this.hookArgsWithOriginalResourcesAppDir)
   }
 
   async buildApp () {
@@ -116,20 +137,15 @@ class App {
   }
 
   async copyTemplate () {
-    const hookArgs = [
-      this.originalResourcesAppDir,
-      this.opts.electronVersion,
-      this.opts.platform,
-      this.opts.arch
-    ]
+    await hooks.promisifyHooks(this.opts.beforeCopy, this.hookArgsWithOriginalResourcesAppDir)
 
     await fs.copy(this.opts.dir, this.originalResourcesAppDir, {
       filter: copyFilter.userPathFilter(this.opts),
       dereference: this.opts.derefSymlinks
     })
-    await hooks.promisifyHooks(this.opts.afterCopy, hookArgs)
+    await hooks.promisifyHooks(this.opts.afterCopy, this.hookArgsWithOriginalResourcesAppDir)
     if (this.opts.prune) {
-      await hooks.promisifyHooks(this.opts.afterPrune, hookArgs)
+      await hooks.promisifyHooks(this.opts.afterPrune, this.hookArgsWithOriginalResourcesAppDir)
     }
   }
 
@@ -171,7 +187,7 @@ class App {
       common.warning('prebuiltAsar has been specified, all asar options will be ignored')
     }
 
-    for (const hookName of ['afterCopy', 'afterPrune']) {
+    for (const hookName of ['beforeCopy', 'afterCopy', 'afterPrune']) {
       if (this.opts[hookName]) {
         throw new Error(`${hookName} is incompatible with prebuiltAsar`)
       }
@@ -202,6 +218,9 @@ class App {
     }
 
     debug(`Running asar with the options ${JSON.stringify(this.asarOptions)}`)
+
+    await hooks.promisifyHooks(this.opts.beforeAsar, this.hookArgsWithOriginalResourcesAppDir)
+
     await asar.createPackageWithOptions(this.originalResourcesAppDir, this.appAsarPath, this.asarOptions)
     const { headerString } = asar.getRawHeader(this.appAsarPath)
     this.asarIntegrity = {
@@ -211,6 +230,8 @@ class App {
       }
     }
     await fs.remove(this.originalResourcesAppDir)
+
+    await hooks.promisifyHooks(this.opts.afterAsar, this.hookArgsWithOriginalResourcesAppDir)
   }
 
   async copyExtraResources () {
@@ -218,9 +239,18 @@ class App {
 
     const extraResources = common.ensureArray(this.opts.extraResource)
 
+    const hookArgs = [
+      this.stagingPath,
+      ...this.commonHookArgs
+    ]
+
+    await hooks.promisifyHooks(this.opts.beforeCopyExtraResources, hookArgs)
+
     await Promise.all(extraResources.map(
       resource => fs.copy(resource, path.resolve(this.stagingPath, this.resourcesDir, path.basename(resource)))
     ))
+
+    await hooks.promisifyHooks(this.opts.afterCopyExtraResources, hookArgs)
   }
 
   async move () {
@@ -229,6 +259,15 @@ class App {
     if (this.opts.tmpdir !== false) {
       debug(`Moving ${this.stagingPath} to ${finalPath}`)
       await fs.move(this.stagingPath, finalPath)
+    }
+
+    if (this.opts.afterComplete) {
+      const hookArgs = [
+        finalPath,
+        ...this.commonHookArgs
+      ]
+
+      await hooks.promisifyHooks(this.opts.afterComplete, hookArgs)
     }
 
     return finalPath